clean up utils structure
