+++ src/kulprit/data/__init__.py
@@ -1,0 +1,3 @@
+"""Model data module."""
+
+from .data import ModelData
